+++ src/MiloneSyntax/Typing.fs
@@ -474,7 +474,7 @@ let private tyIsPtr ty =
   match tk with
   | VoidPtrTk _
   | NativePtrTk _
-  | NativeFunTk -> true
+  | FunPtrTk -> true
   | _ -> false
 
 let private addTraitBound theTrait loc (ctx: TyCtx) =
@@ -545,7 +545,7 @@ let private resolveTraitBound (ctx: TyCtx) theTrait loc : TyCtx =
       | _ when tyIsBasic ty -> true, memo
 
       | TupleTk, []
-      | NativeFunTk, _ -> true, memo
+      | FunPtrTk, _ -> true, memo
 
       // Don't memoize structural types
       // because it doesn't cause infinite recursion
@@ -666,7 +666,7 @@ let private resolveTraitBound (ctx: TyCtx) theTrait loc : TyCtx =
     | ListTk
     | VoidPtrTk _
     | NativePtrTk _
-    | NativeFunTk -> ok ctx
+    | FunPtrTk -> ok ctx
 
     | _ -> error ctx
 
@@ -873,8 +873,14 @@ let private instantiateVariant variantSerial loc (ctx: TyCtx) : Ty * Ty * Ty * T
     let unionTy = tyAssign assignment unionTy
     payloadTy, unionTy, variantTy, ctx
 
-let private castFunAsNativeFun funSerial (ctx: TyCtx) : Ty * TyCtx =
+let private castFunAsNativeFun funSerial loc (ctx: TyCtx) : Ty * TyCtx =
   let funDef = ctx.Funs |> mapFind funSerial
+
+  let ctx =
+    if funDef.Nonlocal then
+      addError ctx "Pointer to a local function is unavailable since it might capture local variables." loc
+    else
+      ctx
 
   // Mark this function as extern "C".
   let ctx =
@@ -1373,7 +1379,8 @@ let private inferPrimExpr ctx prim loc =
   | TPrim.PtrRead
   | TPrim.PtrWrite
   | TPrim.PtrAsIn
-  | TPrim.PtrAsNative -> errorExpr ctx "This function misses some argument." loc
+  | TPrim.PtrAsNative
+  | TPrim.FunPtrInvoke -> errorExpr ctx "This function misses some argument." loc
 
 let private inferRecordExpr ctx expectOpt baseOpt fields loc =
   // First, infer base if exists.
@@ -1696,11 +1703,6 @@ let private inferPrimAppExpr ctx itself =
 
     TNodeExpr(TPtrReadEN, [ expr ], itemTy, loc), itemTy, ctx
 
-  // __nativeFun f
-  | TPrim.NativeFun, TFunExpr (funSerial, _, _) ->
-    let targetTy, ctx = castFunAsNativeFun funSerial ctx
-    TNodeExpr(TNativeFunEN funSerial, [], targetTy, loc), targetTy, ctx
-
   // __nativeFun "funName"
   | TPrim.NativeFun, TLitExpr (StringLit funName, _) ->
     let targetTy, ctx = ctx |> freshMetaTyForExpr itself
@@ -1748,7 +1750,7 @@ let private inferPrimAppExpr ctx itself =
                | TLitExpr _
                | TVarExpr _ -> inferExpr ctx None arg
 
-               | TNodeExpr (TAppEN, [ TPrimExpr (TPrim.NativeFun, _, _); TFunExpr _ ], _, _) -> inferExpr ctx None arg
+               | TNodeExpr (TPtrOfEN, [ TFunExpr _ ], _, _) -> inferExpr ctx None arg
 
                | TNodeExpr (TTyPlaceholderEN, [], ty, loc) ->
                  let ty, ctx = resolveAscriptionTy ctx ty
@@ -1786,6 +1788,34 @@ let private inferWriteExpr ctx expr : TExpr * Ty * TyCtx =
   let ctx = unifyTy ctx loc actualItemTy itemTy
   TNodeExpr(TPtrWriteEN, [ ptr; item ], tyUnit, loc), tyUnit, ctx
 
+let private inferFunPtrInvokeExpr ctx expr : TExpr * Ty * TyCtx =
+  let callee, arg, loc =
+    match expr with
+    | TNodeExpr (TAppEN, [ TNodeExpr (TAppEN, [ TPrimExpr (TPrim.FunPtrInvoke, _, loc); callee ], _, _); arg ], _, _) ->
+      callee, arg, loc
+    | _ -> unreachable ()
+
+  let callee, calleeTy, ctx = inferExpr ctx None callee
+
+  let argTys, resultTy =
+    match substTy ctx calleeTy with
+    | Ty (FunPtrTk, tyArgs) ->
+      match splitLast tyArgs with
+      | Some (argTys, resultTy) -> argTys, resultTy
+      | None -> unreachable ()
+    | _ -> unreachable ()
+
+  let argTy =
+    match argTys with
+    | [ ty ] -> ty
+    | [] -> tyUnit
+    | _ -> tyTuple argTys
+
+  let arg, actualArgTy, ctx = inferExpr ctx (Some argTy) arg
+  let ctx = unifyTy ctx loc actualArgTy argTy
+
+  TNodeExpr(TFunPtrInvokeEN, [ callee; arg ], resultTy, loc), resultTy, ctx
+
 let private inferMinusExpr ctx arg loc =
   let arg, argTy, ctx = inferExpr ctx None arg
 
@@ -1800,6 +1830,11 @@ let private inferPtrOfExpr ctx arg loc =
     let arg, argTy, ctx = inferExpr ctx None arg
     let ty = tyInPtr argTy
     TNodeExpr(TPtrOfEN, [ arg ], ty, loc), ty, ctx
+
+  | TFunExpr (funSerial, _, _) ->
+    let arg, _, ctx = inferExpr ctx None arg
+    let ty, ctx = castFunAsNativeFun funSerial loc ctx
+    TNodeExpr(TFunPtrOfEN, [ arg ], ty, loc), ty, ctx
 
   | _ ->
     let ctx = addError ctx "Expected a variable." loc
@@ -1865,6 +1900,8 @@ let private inferNodeExpr ctx expr : TExpr * Ty * TyCtx =
   match kind, args with
   | TAppEN, [ TPrimExpr _; _ ] -> inferPrimAppExpr ctx expr
   | TAppEN, [ TNodeExpr (TAppEN, [ TPrimExpr (TPrim.PtrWrite, _, _); _ ], _, _); _ ] -> inferWriteExpr ctx expr
+  | TAppEN, [ TNodeExpr (TAppEN, [ TPrimExpr (TPrim.FunPtrInvoke, _, _); _ ], _, _); _ ] ->
+    inferFunPtrInvokeExpr ctx expr
   | TAppEN, [ _; _ ] -> inferAppExpr ctx expr
   | TAppEN, _ -> unreachable ()
 
@@ -1894,12 +1931,13 @@ let private inferNodeExpr ctx expr : TExpr * Ty * TyCtx =
 
   | TAbortEN, _ -> txAbort ctx loc
 
+  | TFunPtrOfEN, _
   | TDiscriminantEN _, _
   | TCallNativeEN _, _
-  | TPtrOffsetEN _, _
-  | TPtrReadEN _, _
-  | TPtrWriteEN _, _
-  | TNativeFunEN _, _
+  | TPtrOffsetEN, _
+  | TPtrReadEN, _
+  | TPtrWriteEN, _
+  | TFunPtrInvokeEN, _
   | TNativeExprEN _, _
   | TNativeStmtEN _, _
   | TNativeDeclEN _, _ -> unreachable ()